+++ tools/test-js-with-node
@@ -65,6 +65,7 @@ EXEMPT_FILES = make_set(
         "web/src/compose_banner.ts",
         "web/src/compose_closed_ui.js",
         "web/src/compose_fade.js",
+        "web/src/compose_recipient.js",
         "web/src/compose_state.js",
         "web/src/compose_ui.js",
         "web/src/compose_validate.js",
