+++ web/src/narrow.js
@@ -7,6 +7,7 @@ import * as compose_actions from "./compose_actions";
 import * as compose_banner from "./compose_banner";
 import * as compose_closed_ui from "./compose_closed_ui";
 import * as compose_fade from "./compose_fade";
+import * as compose_recipient from "./compose_recipient";
 import * as compose_state from "./compose_state";
 import * as condense from "./condense";
 import {Filter} from "./filter";
@@ -183,7 +184,7 @@ export function reset_ui_state() {
 
 export function handle_middle_pane_transition() {
     if (compose_state.composing) {
-        compose_actions.update_narrow_to_recipient_visibility();
+        compose_recipient.update_narrow_to_recipient_visibility();
     }
 }
