+++ web/src/ui_init.js
@@ -23,7 +23,7 @@ import * as common from "./common";
 import * as compose from "./compose";
 import * as compose_closed_ui from "./compose_closed_ui";
 import * as compose_pm_pill from "./compose_pm_pill";
-import * as compose_ui from "./compose_ui";
+import * as compose_recipient from "./compose_recipient";
 import * as composebox_typeahead from "./composebox_typeahead";
 import * as condense from "./condense";
 import * as copy_and_paste from "./copy_and_paste";
@@ -626,14 +626,12 @@ export function initialize_everything() {
     scroll_bar.initialize();
     message_viewport.initialize();
     navbar_alerts.initialize();
-    compose_closed_ui.initialize();
     initialize_kitchen_sink_stuff();
     echo.initialize();
     stream_edit.initialize();
     user_group_edit.initialize();
     stream_edit_subscribers.initialize();
     stream_data.initialize(stream_data_params);
-    compose_ui.initialize_compose_stream_dropdown();
     user_group_edit_members.initialize();
     pm_conversations.recent.initialize(pm_conversations_params);
     user_topics.initialize();
@@ -653,7 +651,9 @@ export function initialize_everything() {
     message_view_header.initialize();
     server_events.initialize();
     user_status.initialize(user_status_params);
+    compose_recipient.initialize();
     compose_pm_pill.initialize();
+    compose_closed_ui.initialize();
     search_pill_widget.initialize();
     reload.initialize();
     user_groups.initialize(user_groups_params);
